agents/agents/code-reviewer/evals/tasks/python-security/fixture/requirements.txt
flask==3.1.0
PyJWT==2.9.0
